feat：增加职位类型接口

+++ ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/dm/OpeSysPosition.java
@@ -99,7 +99,7 @@ public class OpeSysPosition implements Serializable {
      */
     @TableField(value = "updated_by")
     @ApiModelProperty(value = "更新人")
-    private Integer updatedBy;
+    private long updatedBy;
 
     /**
      * 更新时间
